refactor: remove feature files from app navigator directory and move to a features directory
