+++ app/(tabs)/favourites/index.tsx
@@ -1,3 +1,3 @@
-import { FavouritesLoader } from "./Favourites.loader";
+import { FavouritesLoader } from "../../../src/features/Favourites/Favourites.loader";
 
 export default FavouritesLoader;
