+++ app/launchDetails/[flightNumber].ts
@@ -1,3 +1,3 @@
-import { LaunchDetailsLoader } from "./LaunchDetails.loader";
+import { LaunchDetailsLoader } from "../../src/features/LaunchDetails/LaunchDetails.loader";
 
 export default LaunchDetailsLoader;
