+++ src/features/LaunchDetails/types.ts
@@ -1,4 +1,4 @@
-import { Launch } from "../../src/services/types";
+import { Launch } from "../../services/types";
 
 export interface LaunchDetailsProps {
   launch: Launch;
