+++ src/features/Launches/Launches.tsx
@@ -2,7 +2,7 @@ import { View } from "react-native";
 import { styles } from "./styles";
 import { router } from "expo-router";
 import { LaunchesProps } from "./types";
-import { LaunchesList } from "../../../src/components/LaunchesList";
+import { LaunchesList } from "../../components/LaunchesList";
 
 export const Launches = ({ launches }: LaunchesProps) => (
   <View style={styles.container}>
