+++ src/features/Launches/__tests__/Launches.loader.test.tsx
@@ -1,17 +1,10 @@
 import { http, HttpResponse, delay } from "msw";
 import { setupServer } from "msw/node";
 import { RenderAPI, fireEvent, waitFor } from "@testing-library/react-native";
-import { renderWithProviders } from "../../../src/utils/tests/renderWithProviders";
-import { launchesMock } from "../../../src/__mocks__/launches";
-import { store } from "../../../src/state/store";
-import { LaunchDetailsLoader } from "../LaunchDetails.loader";
-import * as expoRouter from "expo-router";
-
-const mockFlightNumberParam = (flightNumber: string) => {
-  jest.spyOn(expoRouter, "useLocalSearchParams").mockReturnValue({
-    flightNumber,
-  });
-};
+import { renderWithProviders } from "../../../utils/tests/renderWithProviders";
+import { launchesMock } from "../../../__mocks__/launches";
+import { store } from "../../../state/store";
+import { LaunchesLoader } from "../Launches.loader";
 
 const waitForLoadingToFinish = async (tree: RenderAPI) => {
   await waitFor(() => {
@@ -21,42 +14,36 @@ const waitForLoadingToFinish = async (tree: RenderAPI) => {
 
 export const handlers = [
   http.get(
-    "https://api.spacexdata.com/v3/launches/107",
+    "https://api.spacexdata.com/v3/launches/past",
     async () => {
       await delay(150);
       return Response.error();
     },
     { once: true }
   ),
-  http.get("https://api.spacexdata.com/v3/launches/107", async () => {
+  http.get("https://api.spacexdata.com/v3/launches/past", async () => {
     await delay(150);
-    return HttpResponse.json(launchesMock[0]);
+    return HttpResponse.json(launchesMock);
   }),
 ];
 
 const server = setupServer(...handlers);
 
-describe("LaunchDetailsLoader", () => {
+describe("LaunchesLoader", () => {
   let tree: RenderAPI;
 
-  beforeAll(() => {
-    jest.useFakeTimers();
-    server.listen();
-  });
+  beforeAll(() => server.listen());
 
   beforeEach(() => {
     server.resetHandlers();
-    mockFlightNumberParam("107");
-    tree = renderWithProviders(<LaunchDetailsLoader />, { store });
+    tree = renderWithProviders(<LaunchesLoader />, { store });
   });
 
   afterEach(() => {
     jest.clearAllMocks();
   });
 
-  afterAll(() => {
-    server.close();
-  });
+  afterAll(() => server.close());
 
   describe("when the request fails and the user clicks on retry", () => {
     it("retries the request", async () => {
@@ -69,18 +56,22 @@ describe("LaunchDetailsLoader", () => {
       await waitFor(() => {
         expect(tree.getByText("Loading")).toBeDefined();
       });
+
+      await waitForLoadingToFinish(tree);
+
+      expect(tree.getByText("Crew-1")).toBeDefined();
+      expect(tree.getByText("Crew-0")).toBeDefined();
+      expect(tree.getByText("November 16, 2019 1:27 AM")).toBeDefined();
+      expect(tree.getByText("November 16, 2020 1:27 AM")).toBeDefined();
     });
   });
 
-  it("renders the launch details when data is fetched", async () => {
+  it("shows a loading and then renders the launches list when data is fetched", async () => {
     await waitForLoadingToFinish(tree);
 
     expect(tree.getByText("Crew-1")).toBeDefined();
+    expect(tree.getByText("Crew-0")).toBeDefined();
+    expect(tree.getByText("November 16, 2019 1:27 AM")).toBeDefined();
     expect(tree.getByText("November 16, 2020 1:27 AM")).toBeDefined();
-    expect(
-      tree.getByText("Kennedy Space Center Historic Launch Complex 39A")
-    ).toBeDefined();
-    expect(tree.getByText("Falcon 9")).toBeDefined();
-    expect(tree.getByText("SpaceX")).toBeDefined();
   });
 });
